requirements/dev.txt
-r prod.txt

pytest==3.5.1
pytest-flask==0.10.0